Refactor gas handling in VM and AST execution
- Changed return types from (u32, Vec<u8>) to (i64, Vec<u8>) in multiple functions to accommodate larger gas values.
- Updated gas accumulation logic to use i64 for gas tracking across various components including action hooks, execution, and host interactions.
- Introduced a new documentation file outlining the AST/VM return-gas cost model, detailing charging architecture, transaction loop rules, and anti-abuse measures.
- Ensured compatibility with historical transaction charging behavior while implementing new gas handling semantics.

+++ protocol/src/action/macro.rs
@@ -97,7 +97,7 @@ macro_rules! action_define {
         }
 
         impl ActExec for $class {
-            fn execute(&$pself, $pctx: &mut dyn Context) -> Ret<(u32, Vec<u8>)> {
+            fn execute(&$pself, $pctx: &mut dyn Context) -> Ret<(i64, Vec<u8>)> {
                 use std::any::Any;
                 if !$pctx.env().chain.fast_sync {
                     check_action_level($pctx.level(), $pself, $pctx.tx().actions())?;
@@ -106,7 +106,7 @@ macro_rules! action_define {
                 // NOTE: burn_90 gas multiplier is handled centrally in ctx_action_call(),
                 // not here, to keep the logic in one place.
                 #[allow(unused_mut)]
-                let mut $pgas: u32 = $pself.size() as u32;
+                let mut $pgas: i64 = $pself.size() as i64;
                 // execute action body
                 let res: Ret<Vec<u8>> = $exec;
                 let res = res?;
