+++ protocol/src/setup/action_hooker.rs
@@ -5,7 +5,7 @@
     Action hooker
 */
 
-pub type FnActionHookFunc = fn(u16, _act: &dyn Any, _ctx: &mut dyn Context, _gas: &mut u32) -> Rerr ;
+pub type FnActionHookFunc = fn(u16, _act: &dyn Any, _ctx: &mut dyn Context, _gas: &mut i64) -> Rerr ;
 
 pub static mut ACTION_HOOK_FUNC: FnActionHookFunc = |_,_,_,_|Ok(());
 
@@ -15,7 +15,7 @@ pub fn action_hooker(f: FnActionHookFunc) {
     }
 }
 
-pub fn do_action_hook(kid: u16, _act: &dyn Any, _ctx: &mut dyn Context, _gas: &mut u32) -> Rerr {
+pub fn do_action_hook(kid: u16, _act: &dyn Any, _ctx: &mut dyn Context, _gas: &mut i64) -> Rerr {
     unsafe {
         ACTION_HOOK_FUNC(kid, _act, _ctx, _gas)
     }
