+++ vm/src/machine/setup.rs
@@ -52,14 +52,14 @@ pub fn setup_vm_run(ctx: &mut dyn Context, ty: u8, mk: u8, cd: std::sync::Arc<[u
 
 pub fn setup_vm_run_and_collect_gas(
     ctx: &mut dyn Context,
-    gas_acc: &mut u32,
+    gas_acc: &mut i64,
     ty: u8,
     mk: u8,
     cd: std::sync::Arc<[u8]>,
     pm: Value,
 ) -> Ret<Value> {
     let (cost, rv) = setup_vm_run(ctx, ty, mk, cd, pm)?;
-    *gas_acc = gas_acc.saturating_add(cost.max(0) as u32);
+    *gas_acc = gas_acc.saturating_add(cost.max(0));
     Ok(rv)
 }
